+++ salesforce/api/metadata.py
@@ -16,7 +16,6 @@ class MetadataApi():
     def __init__(self, settings, **kwargs):
         self.settings = settings
         self.api_version = settings["api_version"]
-        self.username = settings["username"]
         self.session = None
         self.result = None
 
@@ -25,8 +24,8 @@ class MetadataApi():
 
         Arguments:
 
-        * session_id_expired -- Optional; generally, session in globals() is expired, 
-            if INVALID_SESSION_ID appeared in response requested by session in globals(),
+        * session_id_expired -- Optional; generally, session in .config/session.json is expired, 
+            if INVALID_SESSION_ID appeared in response requested by session in session.json,
             we need to call this method with expired session flag again
 
         Returns:
@@ -34,26 +33,12 @@ class MetadataApi():
         * result -- Keep the session info, if `output_session_info` in plugin setting is True, 
             session info will be outputted to console
         """
-        if self.username not in globals() or session_id_expired:
-            result = soap_login(self.settings)
-
-            # If login succeed, display error and return False
-            if not result["success"]:
-                result["Default Project"] = self.settings["default_project"]["project_name"]
-                self.result = result
-                return self.result
-
-            result["headers"] = {
-                "Authorization": "OAuth " + result["session_id"],
-                "Content-Type": "application/json; charset=UTF-8",
-                "Accept": "application/json"
-            }
-            globals()[self.username] = result
-        else:
-            result = globals()[self.username]
+        result = soap_login(self.settings, session_id_expired)
+        if not result["success"]:
+            self.result = result
+            return self.result
 
         self.session = result
-        self.headers = globals()[self.username]["headers"]
         self.instance_url = result["instance_url"]
         self.partner_url = self.instance_url + "/services/Soap/u/%s.0" % self.api_version
         self.metadata_url = self.instance_url + "/services/Soap/m/%s.0" % self.api_version
@@ -80,7 +65,7 @@ class MetadataApi():
             "SOAPAction": '""'
         }
         soap_body = soap_bodies.describe_layout_body.format(
-            session_id=globals()[self.username]["session_id"], 
+            session_id=self.session["session_id"], 
             sobject=sobject, recordtype_id=recordtype_id)
 
         try:
@@ -129,8 +114,7 @@ class MetadataApi():
             "SOAPAction": '""'
         }
 
-        soap_body = soap_bodies.run_all_test.format(
-            session_id=globals()[self.username]["session_id"])
+        soap_body = soap_bodies.run_all_test.format(session_id=self.session["session_id"])
 
         try:
             response = requests.post(self.apex_url, soap_body, verify=False, 
@@ -210,7 +194,7 @@ class MetadataApi():
 
         soap_body = soap_bodies.execute_anonymous_body.format(
             log_levels=log_levels,
-            session_id=globals()[self.username]["session_id"], 
+            session_id=self.session["session_id"],
             apex_string=apex_string)
 
         try:
@@ -281,7 +265,7 @@ class MetadataApi():
             "SOAPAction": '""'
         }
         soap_body = soap_bodies.check_status_body.format(
-            session_id=globals()[self.username]["session_id"],
+            session_id=self.session["session_id"],
             async_process_id=async_process_id)
 
         response = requests.post(self.metadata_url, soap_body, verify=False, 
@@ -325,7 +309,7 @@ class MetadataApi():
             "SOAPAction": '""'
         }
         soap_body = soap_bodies.check_retrieve_status_body.format(
-            session_id=globals()[self.username]["session_id"],
+            session_id=self.session["session_id"],
             async_process_id=async_process_id)
 
         response = requests.post(self.metadata_url, soap_body, 
@@ -384,9 +368,7 @@ class MetadataApi():
         }
 
         # Populate the soap_body with actual session id
-        parsed_soap_body = soap_body.format(
-            globals()[self.username]["session_id"], 
-            self.api_version)
+        parsed_soap_body = soap_body.format(self.session["session_id"], self.api_version)
 
         # [sf:retrieve]
         util.append_message(panel, "[sf:retrieve] Start request for a retrieve...")
@@ -406,6 +388,7 @@ class MetadataApi():
 
         # Check whether session_id is expired
         if "INVALID_SESSION_ID" in response.text:
+            util.append_message(panel, "[sf:retrieve] Session expired, need login again")
             self.login(True)
             return self.retrieve(soap_body)
 
@@ -509,8 +492,7 @@ class MetadataApi():
             "Content-Type": "text/xml;charset=UTF-8",
             "SOAPAction": '""'
         }
-        soap_body = soap_bodies.cancel_deployment.format(
-            globals()[self.username]["session_id"], async_process_id)
+        soap_body = soap_bodies.cancel_deployment.format(self.session["session_id"], async_process_id)
         response = requests.post(self.metadata_url, soap_body, verify=False, 
             headers=headers)
 
@@ -555,7 +537,7 @@ class MetadataApi():
             "SOAPAction": '""'
         }
         soap_body = soap_bodies.check_deploy_status.format(
-            globals()[self.username]["session_id"], async_process_id)
+            self.session["session_id"], async_process_id)
         response = requests.post(self.metadata_url, soap_body, verify=False, 
             headers=headers)
 
@@ -632,7 +614,7 @@ class MetadataApi():
         util.append_message(panel, "[sf:%s] Start request for a deploy..." % deploy_or_validate)
 
         soap_body = soap_bodies.deploy_package.format(
-            globals()[self.username]["session_id"], 
+            self.session["session_id"],
             base64_zip,
             deploy_options["allowMissingFiles"],
             deploy_options["autoUpdatePackage"],
@@ -658,6 +640,7 @@ class MetadataApi():
 
         # Check whether session_id is expired
         if "INVALID_SESSION_ID" in response.text:
+            util.append_message(panel, "[sf:%s] Session expired, need login again" % deploy_or_validate)
             self.login(True)
             return self.deploy(base64_zip)